+++ Chess.View/StatusAndMode/AutoReviewModeView.xaml
@@ -1,9 +1,9 @@
-﻿<UserControl x:Class="Chess.View.Window.AutoReviewModeView"
+﻿<UserControl x:Class="Chess.View.StatusAndMode.AutoReviewModeView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:Chess.View.Window"
+             xmlns:local="clr-namespace:Chess.View.StatusAndMode"
              xmlns:iconPacks="http://metro.mahapps.com/winfx/xaml/iconpacks"
              mc:Ignorable="d" d:DesignHeight="450" d:DesignWidth="800"
              Loaded="UserControl_Loaded" Unloaded="UserControl_Unloaded">
