+++ Chess.View/StatusAndMode/PlayModeView.xaml
@@ -1,9 +1,9 @@
-﻿<UserControl x:Class="Chess.View.Window.PlayModeView"
+﻿<UserControl x:Class="Chess.View.StatusAndMode.PlayModeView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:Chess.View.Window"
+             xmlns:local="clr-namespace:Chess.View.StatusAndMode"
              mc:Ignorable="d" 
              d:DesignHeight="450" d:DesignWidth="800">
     <StackPanel Orientation="Horizontal">
