+++ Chess.View/StatusAndMode/ReviewModeHeaderDisplyView.xaml
@@ -1,9 +1,9 @@
-﻿<UserControl x:Class="Chess.View.Window.ReviewModeHeaderDisplyView"
+﻿<UserControl x:Class="Chess.View.StatusAndMode.ReviewModeHeaderDisplyView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:Chess.View.Window"
+             xmlns:local="clr-namespace:Chess.View.StatusAndMode"
              xmlns:conv="clr-namespace:Chess.View.Converter"
              xmlns:game="clr-namespace:Chess.ViewModel.Game;assembly=Chess.ViewModel"
              mc:Ignorable="d"
