+++ Chess.View/StatusAndMode/ReviewModeHeaderDisplyView.xaml.cs
@@ -1,6 +1,6 @@
 ﻿using System.Windows.Controls;
 
-namespace Chess.View.Window
+namespace Chess.View.StatusAndMode
 {
     /// <summary>
     /// Interaction logic for ReviewModeHeaderDisplyView.xaml
