+++ Chess.View/StatusAndMode/StatusDisplayView.xaml
