+++ Chess.View/StatusAndMode/StatusDisplayView.xaml.cs
